+++ stdlib/Unicode/src/Unicode.jl
@@ -4,14 +4,23 @@ __precompile__(true)
 
 module Unicode
 
-using Base.Unicode: normalize_string, graphemes, is_assigned_char, textwidth, isvalid,
-                    islower, isupper, isalpha, isdigit, isxdigit, isnumber, isalnum,
+using Base.Unicode: normalize, graphemes, isassigned, textwidth, isvalid,
+                    islower, isupper, isalpha, isdigit, isxdigit, isnumeric, isalnum,
                     iscntrl, ispunct, isspace, isprint, isgraph,
                     lowercase, uppercase, titlecase, lcfirst, ucfirst
 
-export normalize_string, graphemes, is_assigned_char, textwidth, isvalid,
-       islower, isupper, isalpha, isdigit, isxdigit, isnumber, isalnum,
+export normalize, graphemes, isassigned, textwidth, isvalid,
+       islower, isupper, isalpha, isdigit, isxdigit, isnumeric, isalnum,
        iscntrl, ispunct, isspace, isprint, isgraph,
        lowercase, uppercase, titlecase, lcfirst, ucfirst
+
+# BEGIN 0.7 deprecations
+
+@deprecate isnumber(c::Char) Unicode.isnumeric(c)
+@deprecate is_assigned_char(c::Char) Unicode.isassigned(c)
+@deprecate normalize_string(s::AbstractString, nf::Symbol; kwargs...) Unicode.normalize(s, nf; kwargs...)
+@deprecate normalize_string(s::AbstractString; kwargs...) Unicode.normalize(s; kwargs...)
+
+# END 0.7 deprecations
 
 end